+++ mc_examples/realistic_workflows/graphene_slitpore/gcmc_bulk/run.py
